refactor grouping card into own file and fix styles
